Move ApplicationFiles to a dedicated file
